+++ lib/comparator.dart
@@ -1,5 +1,6 @@
 /// Provides default [Comparator] constructors, and various static extension
 /// methods to modify and perform common operations on [Comparator] functions.
+export 'src/comparator/compare_operators.dart';
 export 'src/comparator/constructors/explicit.dart';
 export 'src/comparator/constructors/natural.dart';
 export 'src/comparator/constructors/reverse.dart';